MAINT: Renamed `dynamic_types` to `platform_aliases`
