+++ numpy/typing/_platform_aliases.py
