+++ numpy/typing/tests/data/reveal/platform_aliases.py
